remove branch references from nightly-test workflow
I want to try running these builds from a branch other than master.
Workflows with workflow_dispatch let you specify a branch to run on,
scheduled runs run on the default branch anyway, so hardcoding the
branch doesn't seem helpful.

This can be ported to the proper nightly.yml too if it works.