Switch to earlier pyinstaller version
misc/requirements/requirements-pyinstaller.txt-raw
With the latest pyinstaller I'm getting

    [4964] LOADER: Python library: Y:\qutebrowser\python310.dll
    Error loading Python DLL 'Y:\qutebrowser\python310.dll'.
    LoadLibrary: The specified module could not be found.

As of https://github.com/pyinstaller/pyinstaller/commit/896db9ba652585df0f8bec5d944e79b82e5dde58
our smoke tests don't work. After running through the build steps
locally on windows it looks like it can't find the python dll. The
layout is like:

    dist/
      qutebrowser/
        qutebrowser.exe
        _internal/
          all the runtime files etc

Moving qutebrowser.exe into `_internal/` and running it from there seems
to make things work fine.

This change was made for the "onedir" builds. Are we using that? Should
we be? idk

For macOS there is an `_internal/` folder but it's in `qutebrowser/`,
not `qutebrowser.app/`, so I'm not sure if it's the same issue there.
We'll see if the smoke test gets better after this downgrade.