Do raise a smoke test error on debug builds
I don't get it. If you capture the output on a debug build and don't
print it what's the point? And why have --debug args in the workflow if
debug is supposed to be called from a failure in the non debug build? I
feel like I'm missing something here.

Hopefully this lets me actually see some output in CI ...