Change resources lookup based on pyinstaller advice
From https://pyinstaller.org/en/stable/runtime-information.html

They recommend using the path of a python module to get to the bundle
dir, since it should be under it. So go up a fixed amount from this file
and that should do it. Looks like it'll work for the macOS build, not
sure about windows yet.

TODO: clean up the comments :)